chore: 完成多模块本地开发环境适配优化
1. 调整各服务配置适配本地开发环境：修改RabbitMQ、Redis连接地址为localhost，关闭Sentinel和Sleuth，禁用Feign Sentinel支持
2. 修复类型不匹配问题：将购物车相关方法参数从Integer改为Long，清理过期前端代码
3. 优化异常处理和业务逻辑：完善登录回调跳转、购物车合并、库存查询逻辑，修复支付金额获取
4. 更新Nginx配置适配本地容器部署，新增启动脚本JDK版本校验
5. 重构部分页面样式和结构，修复代码空指针风险

+++ gulimall-cart/src/main/java/com/xunqi/gulimall/cart/controller/CartController.java
@@ -128,7 +128,7 @@ public class CartController {
      * @return
      */
     @GetMapping(value = "/deleteItem")
-    public String deleteItem(@RequestParam("skuId") Integer skuId) {
+    public String deleteItem(@RequestParam("skuId") Long skuId) {
 
         cartService.deleteIdCartInfo(skuId);
